+++ packages/gizmo/src/Utils.ts
@@ -1,7 +1,7 @@
-import { Material, Mesh, Layer, Engine, Vector4, PrimitiveMesh, Vector3, BlinnPhongMaterial } from "oasis-engine";
-
-import { LinesMesh, ArcLineMesh } from "./Mesh";
-import { createMeshMaterial } from "./material";
+import { Material, Mesh, Engine, Vector4, PrimitiveMesh } from "oasis-engine";
+import { ArcLineMesh } from "./ArcLineMesh";
+import { LinesMesh } from "./LineMesh";
+import { createMeshMaterial } from "./MeshMaterial";
 
 class Utils {
   public redMaterial: Material;
