+++ packages/gizmo/src/enums/GizmoState.ts
@@ -1,0 +1,17 @@
+/**
+ * Gizmo State
+ */
+export enum GizmoState {
+  /**
+   * translate state
+   */
+  translate = "translate",
+  /**
+   * rotate state
+   */
+  rotate = "rotate",
+  /**
+   * scale state
+   */
+  scale = "scale"
+}
