+++ packages/gizmo/src/index.ts
@@ -1,1 +1,2 @@
 export { GizmoControls } from "./GizmoControls";
+export { GizmoState } from './enums/GizmoState'
